convert cancelled-timer-in-callback
